requirements.txt
pandas==2.2.3
numpy==2.2.2
scipy==1.15.1
pyyaml==6.0.2
matplotlib==3.10.0
scikit-learn==1.6.1
streamlit==1.42.0
plotly==6.0.0
